Ensure 0-length video segments are excluded
This happens adhoc and without clear reason. It seems that the Video AI
API sometimes just returns those 0-length segments.

Because we're looping over DataFrame rows and using the index as the
visual_shot_id we could theoretically get non-continuous IDs
(i.e. 1, 3, 4...). This shouldn't be an issue though as the final A/V
segment ID is what's important for the frontend / users. Also, we've
observed that when this problem happens it ALWAYS happens for the last
shot in the list.

Change-Id: Ica2735df8846c23e783d1da3c3b26bb344572f54