Implement crop-area previews for vertical and square formats.
Including rendering, which is a multi-stage process (first, crop the original video into square / vertical, then select the individual scenes).
This provides smoother cuts than cutting and cropping in one-go (as you'd need to have explicit crop commands that cover every potential frame, and frame timestamps must match those that are generated by the Video AI API, which isn't always the case).
This has the added benefit of generating square and vertical variants of the "input" video, which is a highly requested feature.

Change-Id: Idbb96cbe36de014e789fe5673c17582da1a7c1e9