Add GCS folder and 'share rendered videos' links
The 'share' link is built by setting the value of a URL parameter "inputCombosFolder" to the the combos folder path on GCS (<root-video-folder>/<timestamp>-combos).
Angular's ActivatedRoute is relevant for when the app is served locally (`ng serve`) in order to capture URL parameters, while a hidden input element with the id `input-combos-folder` is appended to the rendered HTML content for the Apps Script served Web App.

The behavior of "rendering" in the UI has also been changed so that the "rendered videos" panel is opened directly after closing the "render queue" sidenav (or opening a "share" URL), and a MatSpinner is shown until the videos are available.

This change also includes a small fix to the "loadVariant" method (clicking on a variant within the render queue) to reset the "played" status of individual segments before previewing.

Change-Id: Ia0551bf490f5a9c6acf0e7658065216b8d653ebd