Disable generation of formats and crop area previews for non landscape input
This also fixes a few issues related to object tracking and smart framing for videos with a higher resolution than 1280x720. The canvas upon which the frames are drawn is maxed out at 1280x720, so all coordinates have to be scaled accordingly (however, rendering crop commands remain the same so as to still render >1280x720 resolutions).

Change-Id: Ib90f96eb7f6044f2ced8eb9a80d2fdab84fd529d