Improve UI error handling
Browser logs are now going to show both Apps Script and Cloud errors as well, and fail as soon as possible for unexpected errors.
In addition, UI timeouts now behave according to the duration of the input video. A 10 minute time out is added per 1 minute of video, up to a max timeout of 1h.

Change-Id: Iae2f471ad97d7cf05a4167f82bcf2e8aa239db3c