Fix mismatches between numeric and string IDs for segments that have already been analysed and cut
During the initial extraction, segments are cut and IDd as '1', '2', ...
After splits, those would change to "1.0", "2.0", ... due to the introduction of n.1, n.2 IDs for split segments, causing their associated resources (1.mp4, 1.png) to no longer be referenced properly.

This fix unifies the behavior for new and existing videos by ensuring that "whole" IDs do not have a n.0 suffix.

Change-Id: I3baeb7f74fbf03a827ba5e24a5bafc84361a16d5